Dataset refactoring
Extracted code from glue to glue_column_handler
Added handlers importing for datasets
