+++ backend/dataall/db/api/redshift_cluster.py
@@ -3,13 +3,13 @@ import logging
 from sqlalchemy import and_, or_, literal
 
 from .. import models, api, exceptions, paginate, permissions
-from . import has_resource_perm, ResourcePolicy, DatasetTable, Environment, Dataset
-from ..models.Enums import ShareItemStatus
+from . import has_resource_perm, ResourcePolicy, Environment, Dataset
 from ...utils.naming_convention import (
     NamingConventionService,
     NamingConventionPattern,
 )
 from ...utils.slugify import slugify
+from dataall.modules.datasets.services.dataset_table import DatasetTableService
 
 log = logging.getLogger(__name__)
 
@@ -495,7 +495,7 @@ class RedshiftCluster:
         session, username, groups, uri, data=None, check_perm=True
     ) -> models.RedshiftClusterDatasetTable:
         cluster = RedshiftCluster.get_redshift_cluster_by_uri(session, uri)
-        table = DatasetTable.get_dataset_table_by_uri(session, data['tableUri'])
+        table = DatasetTableService.get_dataset_table_by_uri(session, data['tableUri'])
         table = models.RedshiftClusterDatasetTable(
             clusterUri=uri,
             datasetUri=data['datasetUri'],
