+++ backend/dataall/modules/datasets/__init__.py
@@ -1,12 +1,14 @@
 """Contains the code related to datasets"""
 import logging
+from typing import List
+
 from dataall.modules.loader import ModuleInterface, ImportMode
 
 log = logging.getLogger(__name__)
 
 
 class DatasetApiModuleInterface(ModuleInterface):
-    """Implements ModuleInterface for notebook GraphQl lambda"""
+    """Implements ModuleInterface for dataset GraphQl lambda"""
 
     @classmethod
     def is_supported(cls, modes):
@@ -16,3 +18,14 @@ class DatasetApiModuleInterface(ModuleInterface):
         import dataall.modules.datasets.api
         log.info("API of datasets has been imported")
 
+
+class DatasetAsyncHandlersModuleInterface(ModuleInterface):
+    """Implements ModuleInterface for dataset async lambda"""
+
+    @classmethod
+    def is_supported(cls, modes: List[ImportMode]):
+        return ImportMode.HANDLERS in modes
+
+    def __init__(self):
+        import dataall.modules.datasets.handlers
+        log.info("Dataset handlers have been imported")
