+++ backend/dataall/modules/datasets/handlers/__init__.py
@@ -1,0 +1,8 @@
+"""
+Contains code with the handlers that are need for async
+processing in a separate lambda function
+"""
+from dataall.modules.datasets.handlers import glue_column_handler
+
+__all__ = ["glue_column_handler"]
+
